Refactor reservation components: remove unused login options, implement payment and reservation summary, add check-in panel and success screens, and enhance ticket display functionality.
