+++ app/reservation/page.tsx
@@ -2,97 +2,51 @@
 
 import {  useEffect } from "react";
 import { useRouter } from "next/navigation";
-import { useReadProfile } from "@/hooks/auth";
 import 'react-day-picker/dist/style.css';
 import { useForm } from "react-hook-form";
-import DateReservationComponent from "@/components/dateReservationComponent";
-import WelcomebackComponent from "@/components/welcomebackComponent";
-import LoginLandingPageComponent from "@/components/loginLandingComponent";
-import LoginOptionsComponent from "@/components/loginOptionsComponent";
-import TransactionSuccessComponent from "@/components/transactionSuccessComponent";
-import SummaryComponent from "@/components/summaryComponent";
-import SeatSelectionComponent from "@/components/seatSelection";
-import PaymentOptionComponent from "@/components/paymentOption";
-import PaymentGatewayComponent from "@/components/paymentGateway";
-import TicketComponent from "@/components/ticket";
+import { useConvexAuth } from "convex/react";
 
 
 export default function Reservation() {
+    const { isAuthenticated } = useConvexAuth();
     const router = useRouter();
-    const profile = useReadProfile();
-
 
     const { register, handleSubmit, watch, formState: { errors } } = useForm<FormData>();
 
-  useEffect(() => {
-    if (!profile) return;
-    // If user already has a profile, redirect to home
-    if (profile?.firstName) {
-      router.push("/");
-    }
-  }, [profile, router]);
-
-  interface FormData {
-    email: string;
-    password: string;
-    seats: string
-  }
+    useEffect(() => {
+      // If user is not authenticated, redirect to sign-in page
+      if (!isAuthenticated) {
+        router.push("/userLogin");
+      }
+    }, [isAuthenticated, router])
 
   const onSubmit = () =>{
 
   }
   
   return (
-    <section className="w-full h-screen flex justify-center items-center relative px-5 xl:p-0">
+    <section className="w-full h-screen flex justify-center items-center p-4 xl:p-0">
 
-      {/* reservation
-      <div className="w-full sm:w-[335px] h-[812px] flex flex-col justify-center items-center">
-        <header className="flex flex-col justify-center items-center">
-          <h3 className="text-xl">Welcome to</h3>
-          <h1 className="text-5xl font-bold my-2">INSPACE</h1>
-          <span className="my-4">Please log in to get started</span>
-        </header>
-          
-      <form onSubmit={handleSubmit(onSubmit)} className="my-4">
+      {/* login options */}
+      <div className="w-full sm:w-[335px] h-[812px] flex flex-col gap-8 justify-center items-center">
+        {/* header text */}
+          <header className="flex flex-col justify-center items-center">
+            <h3 className="text-xl">Welcome to</h3>
+            <h1 className="text-5xl font-bold my-2">INSPACE</h1>
+          </header>
 
-          <input 
-            placeholder="Enter your email" 
-            {...register("email", { required: true })}  
-            className="w-full sm:max-w[335px] h-8 text-xs border p-2 mb-4 rounded-sm "
-          />
+        {/* login options */}
+        <div className="w-full h-fit  flex  flex-col mt-32">
+          {/* guest button */}
+          <a href="/guest" className="w-full h-8 text-xs font-semibold border-2 flex justify-center items-center border-(--button-gray) bg-white text-black hover:bg-(--button-gray) rounded-sm mb-4">Continue as a Guest</a>
+          {/* user button */}
+          <a href="/userLogin" className="w-full h-8 text-xs font-semibold flex justify-center items-center  bg-(--button-gray) text-black hover:bg-gray-300 rounded-sm">Continue as a User</a>
+        </div>
 
-          {errors.email && <span className="text-red-500 text-xs">This field is required</span>}
-
-          <input 
-            placeholder="Enter your password" 
-            {...register("password", { required: true })} 
-            className="w-full sm:max-w[335px] h-8 text-xs border p-2 mb-4 rounded-sm " 
-          />
-
-          {errors.password && <span className="text-red-500 text-xs">This field is required</span>}
-          
-          <button type="submit" className="w-full h-8 text-xs font-semibold bg-(--button-gray) text-black hover:bg-gray-300 rounded-sm">Log In</button>
-
-      </form>
-
-      <div>
-        <span className='text-sm font-normal'>Don&apos;t have an account? </span>
-        <a href="#" className="text-sm font-normal ml-2 underline hover:text-blue-500 cursor-pointer">Register Instead</a>
       </div>
-
-      </div> */}
-
-      {/* <WelcomebackComponent/> */}
-      {/* <DateReservationComponent/> */}
-      {/* <LoginLandingPageComponent/> */}
-      {/* <LoginOptionsComponent/> */}
-      {/* <TransactionSuccessComponent/> */}
-      {/* <SummaryComponent/> */}
-      {/* <TicketComponent/> */}
-      {/* <SeatSelectionComponent/> */}
-      {/* <PaymentOptionComponent/> */}
-      <PaymentGatewayComponent/>
-
     </section>
+
+     
+
   );
 }
