+++ components/seatReservationSuccess.tsx
@@ -1,8 +1,13 @@
 import { Check } from 'lucide-react'
 import React from 'react'
+import ReservationTicketComponent from './reservationTicket';
 
-export default function TransactionSuccessComponent() {
+export default function SeatReservationSuccessComponent() {
+
+    const [isETicket, setIsETicket] = React.useState(false);
+
   return (
+    (isETicket === false) ?
     <div className="w-full sm:w-[335px] h-[812px] flex flex-col justify-center items-center">
         {/* header text */}
         <div className="w-[335px] sm:max-w-[335px] h-[373px] sm:max-h-[373px] flex flex-col justify-center items-center bg-(--background-gray) rounded-lg">
@@ -24,10 +29,12 @@ export default function TransactionSuccessComponent() {
         <button 
             type="button" 
             className="w-full h-8 text-xs font-semibold bg-(--button-gray) text-black hover:bg-gray-300 rounded-sm mt-[140px]"
+            onClick={() => setIsETicket(true)}
         >
             View e-Ticket
         </button>
         
-    </div>
+    </div> : 
+    <ReservationTicketComponent/>
   )
 }
